Update README.md
An active thread on Reddit is \Random Acts of Pizza” where a user can put in a request for a pizza to be donated by another user. Some requests are granted, others are not. The le pizza_requests.txt includes requests from approximately 5600 pizza requests. Data for each request includes a statement describing the nature of the request, and various other pieces of information. The le README_pizza.txt gives more information about the data set.

Use the data set to answer the questions below.

    What proportion of requests were successful? (The requester received pizza.)

    Find the median account age at the time of request for all requests.

    Divide the requests into those with account age greater than the median found in the previous question, and those with account age less than or equal to the median. Find a 95% con dence interval for the